requirements.txt
accelerate==1.2.1
certifi==2023.7.22
charset-normalizer
filelock==3.16.1
fsspec==2024.9.0
huggingface-hub==0.25.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
neo4j
networkx==3.1
numpy==1.26.4
packaging==23.1
peft==0.14.0
psutil==5.9.5
pytz==2023.3.post1
PyYAML==6.0.1
regex
requests==2.32.3
safetensors==0.4.5
setuptools
sympy==1.13.1
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.47.1
trl==0.13.0
triton==3.1.0
typing-extensions==4.12.2
urllib3==1.26.15
pymongo==3.12.3
requests-oauthlib==1.3.1
responses==0.18.0
prometheus-fastapi-instrumentator==7.0.0
prometheus-client==0.21.1
sqlparse==0.2.4
celery==5.4.0
Django==3.2.4
django-celery-beat==2.7.0
django-celery-results==2.4.0
django-extensions==3.2.3
django-prometheus==2.3.1
djangorestframework==3.14.0
djongo==1.3.6
dvc==3.58.0
dvc-s3==3.2.0
redis==4.4.2
wandb==0.19.1